refactor: rename app.routing -> app-routing
For consistency with web Angular.
