+++ src/ng-new/shared/_files/__sourcedir__/app/app-routing.tns.ts
