+++ src/ng-new/shared/_files/__sourcedir__/app/app.module.tns.ts
@@ -1,7 +1,7 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { NativeScriptModule } from 'nativescript-angular/nativescript.module';
 
-import { AppRoutingModule } from './app.routing';
+import { AppRoutingModule } from './app-routing';
 import { AppComponent } from './app.component';
 
 import { BarcelonaModule } from './barcelona/barcelona.module';
